requirements.txt
pytorch
numpy
opencv-python
black
matplotlib
scikit-learn
transformers